image-caption-generator/requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: osx-64
abseil-cpp=20210324.2=he49afe7_0
absl-py=1.0.0=pyhd8ed1ab_0
aiohttp=3.7.0=py39hb5aae12_0
astroid=2.9.3=pypi_0
astunparse=1.6.3=pyhd8ed1ab_0
async-timeout=3.0.1=py_1000
attrs=21.4.0=pyhd8ed1ab_0
blinker=1.4=py_1
bottleneck=1.3.2=py39he3068b8_1
brotli=1.0.9=hb1e8313_2
brotlipy=0.7.0=py39h89e85a6_1003
bzip2=1.0.8=h1de35cc_0
c-ares=1.18.1=hca72f7f_0
ca-certificates=2022.2.1=hecd8cb5_0
cached-property=1.5.2=hd8ed1ab_1
cached_property=1.5.2=pyha770c72_1
cachetools=4.2.4=pyhd8ed1ab_0
cairo=1.16.0=h8023c5d_1
certifi=2021.10.8=py39hecd8cb5_2
cffi=1.15.0=py39hc55c11b_1
chardet=3.0.4=py39h2c36a5b_1008
charset-normalizer=2.0.12=pyhd8ed1ab_0
click=8.0.4=py39h6e9494a_0
cryptography=36.0.1=py39h209aa08_0
cycler=0.11.0=pyhd3eb1b0_0
ffmpeg=4.3.2=h4dad6da_1
fontconfig=2.13.1=ha9ee91d_0
fonttools=4.25.0=pyhd3eb1b0_0
freetype=2.11.0=hd8bbffd_0
future=0.18.2=py39hecd8cb5_1
gast=0.4.0=pyh9f0ad1d_0
gettext=0.21.0=h7535e17_0
giflib=5.2.1=haf1e3a3_0
glib=2.70.2=hcf210ce_4
glib-tools=2.70.2=hcf210ce_4
gmp=6.2.1=h23ab428_2
gnutls=3.6.15=hed9c0bf_0
google-auth=1.35.0=pyh6c4a22f_0
google-auth-oauthlib=0.4.6=pyhd8ed1ab_0
google-pasta=0.2.0=pyh8c360ce_0
graphite2=1.3.14=h38d11af_0
grpc-cpp=1.42.0=h6da9ac5_1
grpcio=1.42.0=py39h757451a_0
h5py=3.6.0=nompi_py39hbc6cb89_100
harfbuzz=3.4.0=h447b35c_0
hdf5=1.12.1=nompi_h2f0ef1a_103
icu=69.1=he49afe7_0
idna=3.3=pyhd8ed1ab_0
importlib-metadata=4.11.3=py39h6e9494a_0
intel-openmp=2022.0.0=hecd8cb5_3615
isort=5.10.1=pypi_0
jasper=2.0.33=h013e400_0
jbig=2.1=h4d881f8_0
jpeg=9e=h0d85af4_0
keras=2.7.0=pyhd8ed1ab_0
keras-preprocessing=1.1.2=pyhd8ed1ab_0
kiwisolver=1.3.2=py39he9d5cce_0
krb5=1.19.2=hcd88c3b_0
lame=3.100=h1de35cc_0
lazy-object-proxy=1.7.1=pypi_0
lcms2=2.12=hf1fd2bf_0
lerc=3.0=he9d5cce_0
libblas=3.9.0=13_osx64_mkl
libcblas=3.9.0=13_osx64_mkl
libcurl=7.81.0=hf45b732_0
libcxx=12.0.0=h2f01273_0
libdeflate=1.8=h9ed2024_5
libedit=3.1.20210910=hca72f7f_0
libev=4.33=h9ed2024_1
libffi=3.4.2=he9d5cce_2
libgfortran=5.0.0=9_3_0_h6c81a4c_23
libgfortran5=9.3.0=h6c81a4c_23
libglib=2.70.2=hf1fb8c0_4
libiconv=1.16=h1de35cc_0
libidn2=2.3.2=h9ed2024_0
liblapack=3.9.0=13_osx64_mkl
liblapacke=3.9.0=13_osx64_mkl
libnghttp2=1.46.0=ha29bfda_0
libopencv=4.5.5=py39h3195029_2
libpng=1.6.37=ha441bb4_0
libprotobuf=3.19.4=hcf210ce_0
libssh2=1.10.0=h52ee1ee_2
libtasn1=4.16.0=h9ed2024_0
libtiff=4.3.0=hd146c10_2
libunistring=0.9.10=h9ed2024_0
libwebp=1.2.2=h56c3ce4_0
libwebp-base=1.2.2=hca72f7f_0
libxml2=2.9.12=h7e28ab6_1
libzlib=1.2.11=h9173be1_1013
llvm-openmp=13.0.1=hcb1a161_1
lz4-c=1.9.3=h23ab428_1
markdown=3.3.6=pyhd8ed1ab_0
matplotlib=3.5.1=py39hecd8cb5_0
matplotlib-base=3.5.1=py39hfb0c5b7_0
mccabe=0.6.1=pypi_0
mkl=2022.0.0=hecd8cb5_105
multidict=6.0.2=py39h89e85a6_0
munkres=1.1.4=py_0
ncurses=6.3=hca72f7f_2
nettle=3.7.3=h230ac6f_1
ninja=1.10.2=py39haf03e11_3
numexpr=2.8.0=py39h4d6be9b_1
numpy=1.21.5=py39h516c64c_0
oauthlib=3.2.0=pyhd8ed1ab_0
opencv=4.5.5=py39h6e9494a_2
openh264=2.1.1=h8346a28_0
openssl=1.1.1m=hca72f7f_0
opt_einsum=3.3.0=pyhd8ed1ab_1
packaging=21.3=pyhd3eb1b0_0
pandas=1.4.1=py39he9d5cce_0
pcre=8.45=h23ab428_0
pillow=9.0.1=py39hde71d04_0
pip=21.2.4=py39hecd8cb5_0
pixman=0.40.0=h9ed2024_1
platformdirs=2.5.1=pypi_0
protobuf=3.19.4=py39h9fcab8e_0
py-opencv=4.5.5=py39h71a6800_2
pyasn1=0.4.8=py_0
pyasn1-modules=0.2.7=py_0
pycparser=2.21=pyhd3eb1b0_0
pyjwt=2.3.0=pyhd8ed1ab_1
pylint=2.12.2=pypi_0
pyopenssl=22.0.0=pyhd8ed1ab_0
pyparsing=3.0.4=pyhd3eb1b0_0
pysocks=1.7.1=py39h6e9494a_4
python=3.9.10=h1dd9edd_2_cpython
python-dateutil=2.8.2=pyhd3eb1b0_0
python-flatbuffers=2.0=pyhd8ed1ab_0
python_abi=3.9=2_cp39
pytorch=1.10.2=cpu_py39h19568cb_1
pytz=2021.3=pyhd3eb1b0_0
pyu2f=0.1.5=pyhd8ed1ab_0
re2=2021.11.01=he49afe7_0
readline=8.1.2=hca72f7f_1
requests=2.27.1=pyhd8ed1ab_0
requests-oauthlib=1.3.1=pyhd8ed1ab_0
rsa=4.8=pyhd8ed1ab_0
scipy=1.8.0=py39h056f1c0_1
setuptools=58.0.4=py39hecd8cb5_0
six=1.16.0=pyhd3eb1b0_1
sleef=3.5.1=h6db0672_2
snappy=1.1.8=hb1e8313_3
sqlite=3.37.2=h707629a_0
tensorboard=2.6.0=pyhd8ed1ab_1
tensorboard-data-server=0.6.0=py39h209aa08_1
tensorboard-plugin-wit=1.8.1=pyhd8ed1ab_0
tensorflow=2.7.0=cpu_py39hb0a6171_0
tensorflow-base=2.7.0=cpu_py39h326c378_0
tensorflow-estimator=2.7.0=cpu_py39hf4c5dbc_0
termcolor=1.1.0=py_2
tk=8.6.11=h7bc2e8c_0
toml=0.10.2=pypi_0
torchvision=0.11.3=py39_cpu
tornado=6.1=py39h9ed2024_0
typing_extensions=3.10.0.2=pyh06a4308_0
tzdata=2021e=hda174b7_0
urllib3=1.26.8=pyhd8ed1ab_1
werkzeug=2.0.3=pyhd8ed1ab_1
wheel=0.37.1=pyhd3eb1b0_0
wrapt=1.13.3=pypi_0
x264=1!161.3030=h0d85af4_1
xz=5.2.5=h1de35cc_0
yarl=1.7.2=py39h89e85a6_1
zipp=3.7.0=pyhd8ed1ab_1
zlib=1.2.11=h9173be1_1013
zstd=1.5.0=hcb37349_1
